requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
betamax==0.8.1
betamax-serializers==0.2.1
bleach==5.0.0
branca==0.4.2
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
Cython==0.29.32
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
Django==3.2.10
django-cors-headers==3.11.0
django-ipware==4.0.2
djangorestframework==3.13.1
entrypoints==0.3
fastjsonschema==2.15.3
ffmpeg-python==0.2.0
filelock==3.4.2
Flask==2.0.2
Flask-RESTful==0.3.9
folium==0.12.1.post1
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
IMDbPY==2021.4.18
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.2.1
jupyterlab-widgets==1.1.0
linecache2==1.0.0
lxml==4.9.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
MouseInfo==0.1.3
mysqlclient==2.1.0
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.21.4
opencv-python==4.5.4.60
outcome==1.1.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
phpserialize==1.3
pickleshare==0.7.5
Pillow==9.0.0
pipenv==2021.11.23
platformdirs==2.4.1
prometheus-client==0.14.1
prompt-toolkit==3.0.24
psutil==5.8.0
psycopg2==2.9.5
pyaes==1.6.1
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
pydub==0.25.1
pyfcm==1.5.4
PyGetWindow==0.0.9
Pygments==2.10.0
pymongo==4.0.1
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyparsing==3.0.8
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.2
python-magic==0.4.24
python-vlc==3.0.12118
pytweening==1.0.4
pytz==2021.3
pywin32==302
pywinpty==2.0.5
pyzmq==22.3.0
qrcode==7.3.1
requests==2.26.0
rsa==4.9
selenium==4.1.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.39
sqlparse==0.4.2
Telethon==1.25.0
terminado==0.13.3
tinycss2==1.1.1
toml==0.10.2
tornado==6.1
tqdm==4.64.0
traceback2==1.4.0
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
ua-parser==0.10.0
urllib3==1.26.7
user-agents==2.2.0
virtualenv==20.13.0
virtualenv-clone==0.5.7
waitress==2.1.2
Wand==0.6.7
wcwidth==0.2.5
webdriver-manager==3.5.2
webdrivermanager==0.10.0
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.6.0
wsproto==1.0.0
